LayoutTests/imported/w3c/web-platform-tests/credential-management/otpcredential-iframe.https-expected.txt


FAIL Test OTPCredential enabled in same origin iframes assert_equals: expected "Pass" but got "Fail"
FAIL OTPCredential enabled in cross origin iframes with permissions policy assert_equals: expected "Pass" but got "Fail"
FAIL OTPCredential disabled in cross origin iframes without permissions policy assert_equals: expected "NotAllowedError" but got "Error"
